backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
tensorflow==2.17.1
opencv-python==4.12.0.88
numpy==1.26.4
pillow==11.3.0
spotipy==2.23.0
bcrypt==4.1.2
pyjwt==2.8.0
flask-jwt-extended==4.6.0